refactor(naming): rename ng2-validators to ngx-validators
BREAKING CHANGE: new npm package
