+++ .gitignore
@@ -52,5 +52,7 @@ examples/node_modules/
 *.ngsummary.json
 *.metadata.json
 *.ngfactory.ts
+bundles/ngx-validators.umd.js
+bundles/ngx-validators.umd.min.js
 bundles/ng2-validators.umd.js
 bundles/ng2-validators.umd.min.js
