+++ examples/.angular-cli.json
@@ -2,40 +2,37 @@
     "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
     "project": {
         "version": "1.0.0-beta.32.3",
-        "name": "ng2-validators examples"
+        "name": "ngx-validators examples"
     },
-    "apps": [
-        {
-            "root": "src",
-            "outDir": "dist",
-            "assets": [
-                "assets",
-                "favicon.ico"
-            ],
-            "index": "index.html",
-            "main": "main.ts",
-            "polyfills": "polyfills.ts",
-            "test": "test.ts",
-            "tsconfig": "tsconfig.json",
-            "prefix": "app",
-            "styles": [
-                "styles.css"
-            ],
-            "scripts": [],
-            "environmentSource": "environments/environment.ts",
-            "environments": {
-                "dev": "environments/environment.ts",
-                "prod": "environments/environment.prod.ts"
-            }
+    "apps": [{
+        "root": "src",
+        "outDir": "dist",
+        "assets": [
+            "assets",
+            "favicon.ico"
+        ],
+        "index": "index.html",
+        "main": "main.ts",
+        "polyfills": "polyfills.ts",
+        "test": "test.ts",
+        "tsconfig": "tsconfig.json",
+        "prefix": "app",
+        "styles": [
+            "styles.css"
+        ],
+        "scripts": [],
+        "environmentSource": "environments/environment.ts",
+        "environments": {
+            "dev": "environments/environment.ts",
+            "prod": "environments/environment.prod.ts"
         }
-    ],
+    }],
     "e2e": {
         "protractor": {
             "config": "./protractor.conf.js"
         }
     },
-    "lint": [
-        {
+    "lint": [{
             "files": "src/**/*.ts",
             "project": "src/tsconfig.json"
         },
