+++ examples/src/app/app.component.html
@@ -1,7 +1,7 @@
 <md-toolbar color="primary">
     <button md-button class="menu-button" (click)="sidenav.toggle()">
     <md-icon>menu</md-icon></button>
-    <span class="home" [routerLink]="['guide']">ng2-validators</span>
+    <span class="home" [routerLink]="['guide']">ngx-validators</span>
 </md-toolbar>
 
 <md-sidenav-container>
@@ -11,8 +11,8 @@
             <a md-list-item [routerLink]="['guide']" routerLinkActive="selected">Guide</a >
             <a md-list-item [routerLink]="['reactive-forms']" routerLinkActive="selected">Reactive forms</a >
             <a md-list-item [routerLink]="['forms']" routerLinkActive="selected">Forms</a >
-            <a md-list-item href="https://github.com/Nightapes/ng2-validators">Github</a >
-            <a md-list-item href="https://github.com/Nightapes/ng2-validators/issues">Github Issues</a >
+            <a md-list-item href="https://github.com/Nightapes/ngx-validators">Github</a >
+            <a md-list-item href="https://github.com/Nightapes/ngx-validators/issues">Github Issues</a >
         </md-nav-list>
   </md-sidenav>
