+++ examples/src/app/reactive-forms/email-validator/email-validator.component.ts
@@ -1,4 +1,4 @@
-import { EmailValidators } from 'ng2-validators';
+import { EmailValidators } from 'ngx-validators';
 import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
 import { Component, OnInit } from '@angular/core';
