+++ examples/src/index.html
@@ -2,17 +2,17 @@
 <html>
 
 <head>
-  <meta charset="utf-8">
-  <title>ng2-validators</title>
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <base href="/">
+    <meta charset="utf-8">
+    <title>ngx-validators</title>
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <base href="/">
 
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" type="image/x-icon" href="favicon.ico">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
 
 <body>
-  <app-root>Loading...</app-root>
+    <app-root>Loading...</app-root>
 </body>
 
 </html>
